test modules split
