+++ test/test_vector.py
@@ -1,5 +1,4 @@
 from vector import Vector
-from point import Point, LineSegment
 
 
 class TestVector:
@@ -15,11 +14,3 @@ class TestVector:
         v2 = Vector(1, 2)
         res = v1 * v2
         assert res == 3
-
-
-class TestLineSegment:
-    def test_vector_property(self):
-        a = Point(1, 1)
-        b = Point(2, 1)
-        vector = LineSegment(a, b).vector
-        assert vector == Vector(1, 0)
